ADD: constants, clear temporary csv files func, step 3 in loadonesitecsv func
