+++ importers/onesite/loader.go
@@ -27,7 +27,7 @@ func Init() {
 	}
 
 	// get path of splitted csv store
-	SplittedCSVStore = path.Join(path.Dir(filename), "/tempCSVs")
+	SplittedCSVStore = path.Join(path.Dir(filename), "/"+splittedCSVStoreName)
 
 	// if splittedcsvstore not exist then create it
 	if _, err := os.Stat(SplittedCSVStore); os.IsNotExist(err) {
@@ -82,91 +82,12 @@ func LoadOneSiteCSV(userSuppliedValues map[string]string) ([]error, string) {
 	// it is helpful while creating unique files
 	currentTimeFormat := currentTime.Format(time.RFC3339Nano)
 
-	// define column fields with order
-	const (
-		Unit            = iota
-		FloorPlan       = iota
-		UnitDesignation = iota
-		Sqft            = iota
-		UnitLeaseStatus = iota
-		Name            = iota
-		PhoneNumber     = iota
-		Email           = iota
-		MoveIn          = iota
-		NoticeForDate   = iota
-		MoveOut         = iota
-		LeaseStart      = iota
-		LeaseEnd        = iota
-		MarketAddl      = iota
-		DepOnHand       = iota
-		Balance         = iota
-		TotalCharges    = iota
-		Rent            = iota
-		WaterReImb      = iota
-		Corp            = iota
-		Discount        = iota
-		Platinum        = iota
-		Tax             = iota
-		ElectricReImb   = iota
-		Fire            = iota
-		ConcSpecl       = iota
-		WashDry         = iota
-		EmplCred        = iota
-		Short           = iota
-		PetFee          = iota
-		TrashReImb      = iota
-		TermFee         = iota
-		LakeView        = iota
-		Utility         = iota
-		Furn            = iota
-		Mtom            = iota
-		Referral        = iota
-	)
-
-	// defined csv columns
-	var csvCols = []rcsv.CSVColumn{
-		{Name: "Unit", Index: Unit},
-		{Name: "FloorPlan", Index: FloorPlan},
-		{Name: "UnitDesignation", Index: UnitDesignation},
-		{Name: "SQFT", Index: Sqft},
-		{Name: "Unit/LeaseStatus", Index: UnitLeaseStatus},
-		{Name: "Name", Index: Name},
-		{Name: "PhoneNumber", Index: PhoneNumber},
-		{Name: "Email", Index: Email},
-		{Name: "Move-In", Index: MoveIn},
-		{Name: "NoticeForDate", Index: NoticeForDate},
-		{Name: "Move-Out", Index: MoveOut},
-		{Name: "LeaseStart", Index: LeaseStart},
-		{Name: "LeaseEnd", Index: LeaseEnd},
-		{Name: "Market+Addl.", Index: MarketAddl},
-		{Name: "DepOnHand", Index: DepOnHand},
-		{Name: "Balance", Index: Balance},
-		{Name: "TotalCharges", Index: TotalCharges},
-		{Name: "RENT", Index: Rent},
-		{Name: "WATERREIMB", Index: WaterReImb},
-		{Name: "CORP", Index: Corp},
-		{Name: "DISCOUNT", Index: Discount},
-		{Name: "Platinum", Index: Platinum},
-		{Name: "TAX", Index: Tax},
-		{Name: "ELECTRICREIMB", Index: ElectricReImb},
-		{Name: "Fire", Index: Fire},
-		{Name: "CONC/SPECL", Index: ConcSpecl},
-		{Name: "WASH/DRY", Index: WashDry},
-		{Name: "EMPLCRED", Index: EmplCred},
-		{Name: "SHORT", Index: Short},
-		{Name: "PETFEE", Index: PetFee},
-		{Name: "TRASHREIMB", Index: TrashReImb},
-		{Name: "TERMFEE", Index: TermFee},
-		{Name: "Lakeview", Index: LakeView},
-		{Name: "UTILITY", Index: Utility},
-		{Name: "FURN", Index: Furn},
-		{Name: "MTOM", Index: Mtom},
-		{Name: "REFERRAL", Index: Referral},
-	}
-
 	// ###################################
 	// # INIT PHASE : LOAD FIELD MAP IN ONESITE MAP #
 	// ###################################
+
+	// csvCols and consts for all onesite csv fields are defined in
+	// constant.go file
 
 	// Onesite csv headers slice and load it from csvCols
 	OneSiteCSVHeaders := []string{}
@@ -401,6 +322,13 @@ func LoadOneSiteCSV(userSuppliedValues map[string]string) ([]error, string) {
 
 	// rtErrors := rcsv.LoadRentableTypesCSV(rentableTypeCSVFile.Name())
 	// errorList = append(errorList, rtErrors...)
+
+	// ##################################
+	// # PHASE 3 : CLEAR THE TEMPORARY CSV FILES #
+	// ##################################
+	ClearSplittedTempCSVFiles(currentTimeFormat)
+
+	// RETURN
 	return errorList, msg
 }
 
@@ -414,7 +342,17 @@ func rrDoLoad(fname string, handler func(string) []error) []error {
 // that created by program temporarily while loading onesite data
 //  and if any error occurs
 func RollBackSplitOperation(timestamp string) {
+	ClearSplittedTempCSVFiles(timestamp)
+}
 
+// ClearSplittedTempCSVFiles func used only to clear
+// temporarily csv files created by program
+func ClearSplittedTempCSVFiles(timestamp string) {
+	for _, v := range prefixCSVFile {
+		fileName := v + timestamp
+		err := os.Remove(fileName)
+		fmt.Println(err)
+	}
 }
 
 // CSVHandler is main function to handle user uploaded
